add dockerfile

+++ src/main/java/com/doku/live/coding/dokutest/repositories/BankRepository.java
@@ -3,5 +3,5 @@ package com.doku.live.coding.dokutest.repositories;
 import com.doku.live.coding.dokutest.entity.Bank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BankRepository extends JpaRepository<Bank, Long> {
+public interface BankRepository extends JpaRepository<Bank, Integer> {
 }
